Move conversation memory schemas into a separate folder
