+++ griptape/schemas/memory/conversation/conversation_memory_schema.py
