+++ griptape/schemas/memory/conversation/run_schema.py
