+++ griptape/schemas/memory/conversation/summary_conversation_memory_schema.py
